Be more careful about what attribute might be typelib id
Since typelib ids are recorded as a GuidAttribute on an assembly, any
guids from other CustomAttribute entries that are not on an assembly and
are not constructed from a MemberRef can be discarded.

MemberRef can be used to filter because GuidAttribute comes from
mscorlib, so unless we're looking at mscorlib itself GuidAttribute will
always be constructed using a MemberRef into another assembly.